# swagger.yaml
openapi: 3.0.0
info:
  title: upworks
  contact: {}
  version: '1.0'
servers:
- url: https://upworks.onrender.com/api/v1
  variables: {}
paths:
...
  /jobs/{job_id}:
    parameters:
      - in: path
        name: job_id
        schema:
          type: string
        required: true
        description: The user ID
    get:
      tags:
      - Jobs
      summary: Get Single Job
      operationId: GetSingleJob
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    patch:
      tags:
      - Jobs
      summary: Update Job
      operationId: UpdateJob
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UpdateJobRequest'
              - example:
                  title: updated value title
                  skills: gis,html
                  numsOfFreelancers: 5
            example:
              title: updated value title
              skills: gis,html
              numsOfFreelancers: 5
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    delete:
      tags:
      - Jobs
      summary: Delete Job
      operationId: DeleteJob
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
  schemas:
...
    UpdateJobRequest:
      title: UpdateJobRequest
      required:
      - title
      - skills
      - numsOfFreelancers
      type: object
      properties:
        title:
          type: string
        skills:
          type: string
        numsOfFreelancers:
          type: integer
          format: int32
      example:
        title: updated value title
        skills: gis,html
        numsOfFreelancers: 5
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...